Ignore invalid messages with TXID 0x0
A sensitive receiver can pick up a sync sequence from noise data, but if
that same noise data results in sensor data and CRC with values of 0x0,
the CRC will still pass. Instead of waiting for all message data to be
received to check if it is 0x0, the state machine can have a shortcut
from the TXID state to the SYNC state. Blocking TXID 0x0 greatly
decreases the probability of noise data both passing sync and CRC
verification.