Provide console output if ManchesterDecoder size is exceeded
If too many bits are shifted into the ManchesterDecoder to fit the
output, data is lost. If the decoder is later used to receive too large
of fields, or is used on a platform where the data types are smaller,
there will now be an error to help with debugging.