Update instructions in readme to not build as root
sudo make install will both compile and install Soapy345 as root if not already compiled. This means the build/ output folder and its contents (.o and executable) would be under the root owner and group and the user would later not be able to delete or overwrite the folder or its contents without elevated permissions. Only installation itself should be done as root.