Compute number of filter taps using Harris approximation
Using the Harris approximation for the number of filter taps, as found
in the book Multirate Signal Processing for Communication Systems, by
Fredric J. Harris, the number of taps can be determined from the desired
filter transition width.
The optimal number of taps is known for 200ksps, so transition width and
attenuation were back-calculated from the number of taps so that those
universal transition widths apply to any selected sample rate, if change
is required in the future. Ideally, each filter's cutoff frequency,
transition width, and attenuation should later be computed from known
signal characteristics instead of listed as magic values.